Fix some libraries and clean up conversions.

+++ src/stdtypes.c
@@ -7,16 +7,16 @@
 #define RAND_MAX_WIDTH IMAX_BITS(RAND_MAX)
 _Static_assert((RAND_MAX & (RAND_MAX + 1u)) == 0, "RAND_MAX not a Mersenne number");
 
-static uint64 rand_uint64(void) {
+static uint64 rand_uint64(void){
   uint64 r = 0;
-  for (uint8 i=0; i<64; i++) {
-    r = r*2 + rand()%2;
+  for(uint8 i=0; i<64; i++){
+    r = r*2 + (uint64)rand()%2;
   }
   return r;
 }
 
 uint64 generate_random_uint(uint64 min, uint64 max){
-    static bool firstTime = true;
-    if(firstTime){ srand(time(NULL)); firstTime = false; }
-    return (rand_uint64() % (max - min)) + min;
+  static bool firstTime = true;
+  if(firstTime){ srand((uint32)time(NULL)); firstTime = false; }
+  return (rand_uint64() % (max - min)) + min;
 }
